CALCITE-1384 fix running various mvn invocations
Fix running mvn clean install site, but ensuring that the
javacc generated test sources directory is excluded from the
source root when compiling (it's added as a source root by the
javacc plugin).

Also change the package of the parser extension testing classes
to make it more clear that they are only for testing (for added
clarity in the exclusions in the pom).